Small fixes and tweaks

+++ source/Log.cpp
@@ -7,16 +7,16 @@ namespace GTA
 
 	void Log::OnStart()
 	{
-		LogToFile("", "##########\n\n\n\n##########");
+		LogToFile("", false, "##########\n\n\n\n##########");
 	}
 
 	void Log::Debug(... array<String ^> ^message)
 	{
-		LogToFile("[DEBUG]", message);
+		LogToFile("[DEBUG]", true, message);
 	}
 	void Log::Error(... array<String ^> ^message)
 	{
-		LogToFile("[ERROR]", message);
+		LogToFile("[ERROR]", true, message);
 	}
 
 	void Log::DeleteOldLogs()
@@ -24,14 +24,16 @@ namespace GTA
 		DateTime now = DateTime::Now;
 		for each(System::String^ path in IO::Directory::GetFiles(IO::Path::GetDirectoryName(Reflection::Assembly::GetExecutingAssembly()->Location), "*.log"))
 		{
-			DateTime logDate = DateTime::Parse(path->Substring(path->IndexOf("-") + 1, path->IndexOf(".log") - (path->IndexOf("-") + 1)));
-			if ((now - logDate).Days >= maxLogAge){
-				IO::File::Delete(path);
+			if (path->Contains(IO::Path::GetFileNameWithoutExtension(Reflection::Assembly::GetExecutingAssembly()->Location))){
+				DateTime logDate = DateTime::Parse(path->Substring(path->IndexOf("-") + 1, path->IndexOf(".log") - (path->IndexOf("-") + 1)));
+				if ((now - logDate).Days >= maxLogAge){
+					IO::File::Delete(path);
+				}
 			}
 		}
 	}
 
-	void Log::LogToFile(String ^logLevel, ... array<String ^> ^message)
+	void Log::LogToFile(String ^logLevel, bool showTimeStamp, ... array<String ^> ^message)
 	{
 		DateTime now = DateTime::Now;
 		String ^logpath = IO::Path::ChangeExtension(Reflection::Assembly::GetExecutingAssembly()->Location, ".log");
@@ -45,7 +47,14 @@ namespace GTA
 
 			try
 			{
-				sw->Write(String::Concat("[", now.ToString("HH:mm:ss"), "] ", logLevel, " "));
+				if (showTimeStamp)
+				{
+					sw->Write(String::Concat("[", now.ToString("HH:mm:ss"), "] ", logLevel, " "));
+				}
+				else
+				{
+					sw->Write(String::Concat(logLevel, " "));
+				}
 
 				for each (String ^string in message)
 				{
